apiauth/requirements.txt
Flask
Flask-RESTful
pyotp
uWSGI
requests
mysqlclient
PyJWT
xmltodict
lxml
Flask-SQLAlchemy
setuptools
setuptools-rust
wheel